factorise out pghkp/types
