+++ src/hockeypuck/pghkp/storage.go
@@ -18,25 +18,17 @@
 package pghkp
 
 import (
-	"bytes"
 	"database/sql"
-	"fmt"
 	"strings"
 	"sync"
-	"time"
-	"unicode"
-	"unicode/utf8"
 
 	"gopkg.in/tomb.v2"
 
 	_ "github.com/lib/pq"
 	"github.com/pkg/errors"
-	"golang.org/x/exp/slices"
 
 	hkpstorage "hockeypuck/hkp/storage"
 	"hockeypuck/openpgp"
-
-	log "github.com/sirupsen/logrus"
 )
 
 const (
@@ -176,255 +168,4 @@ func (st *storage) createIndexes() error {
 		return errors.WithStack(err)
 	}
 	return nil
-}
-
-// keyDoc is a nearly-raw copy of a row in the PostgreSQL `keys` table.
-type keyDoc struct {
-	RFingerprint string
-	CTime        time.Time
-	MTime        time.Time
-	IdxTime      time.Time
-	MD5          string
-	Doc          string
-	Keywords     string
-}
-
-func readOneKey(b []byte, rfingerprint string) (*openpgp.PrimaryKey, error) {
-	kr := openpgp.NewKeyReader(bytes.NewBuffer(b))
-	keys, err := kr.Read()
-	if err != nil {
-		return nil, errors.WithStack(err)
-	}
-	if len(keys) == 0 {
-		return nil, nil
-	} else if len(keys) > 1 {
-		return nil, errors.Errorf("multiple keys in record: %v, %v", keys[0].Fingerprint(), keys[1].Fingerprint())
-	}
-	if keys[0].RFingerprint != rfingerprint {
-		return nil, errors.Errorf("RFingerprint mismatch: expected=%q got=%q",
-			rfingerprint, keys[0].RFingerprint)
-	}
-	return keys[0], nil
-}
-
-func keywordsTSVector(key *openpgp.PrimaryKey) string {
-	keywords, _, _ := keywordsFromKey(key)
-	tsv, err := keywordsToTSVector(keywords, " ")
-	if err != nil {
-		// In this case we've found a key that generated
-		// an invalid tsvector - this is pretty much guaranteed
-		// to be a bogus key, since having a valid key with
-		// user IDs that exceed limits is highly unlikely.
-		// In the future we should catch this earlier and
-		// reject it as a bad key, but for now we just skip
-		// storing keyword information.
-		log.Warningf("keywords for rfp=%q exceeds limit, ignoring: %v", key.RFingerprint, err)
-		return ""
-	}
-	return tsv
-}
-
-func keywordsTSQuery(query string) (string, error) {
-	keywords, _ := keywordsFromSearch(query)
-	tsq, err := keywordsToTSVector(keywords, " & ")
-	if err != nil {
-		log.Warningf("cannot convert search string to tsquery: %v", err)
-		return "", err
-	}
-	return tsq, nil
-}
-
-// A woefully incomplete list of PostgreSQL stop words to minimise TSVector churn.
-// https://github.com/postgres/postgres/blob/master/src/backend/snowball/stopwords/english.stop
-var pgEnglishStopWords = []string{
-	"i", "me", "my", "myself", "we", "our", "ours", "ourselves", "you", "your", "yours", "yourself", "yourselves",
-	"he", "him", "his", "himself", "she", "her", "hers", "herself", "it", "its", "itself",
-	"they", "them", "their", "theirs", "themselves", "what", "which", "who", "whom", "this", "that", "these", "those",
-	"am", "is", "are", "was", "were", "be", "been", "being", "have", "has", "had", "having", "do", "does", "did", "doing",
-	"a", "an", "the", "and", "but", "if", "or",
-	"because", "as", "until", "while", "of", "at", "by", "for", "with", "about", "against", "between", "into",
-	"through", "during", "before", "after", "above", "below",
-	"to", "from", "up", "down", "in", "out", "on", "off", "over", "under", "again", "further",
-	"then", "once", "here", "there", "when", "where", "why", "how", "all", "any", "both", "each",
-	"few", "more", "most", "other", "some", "such",
-	"no", "nor", "not", "only", "own", "same", "so", "than", "too", "very",
-	"s", "t", "can", "will", "just", "don", "should", "now",
-	"[", "\\", "]", "^", "_", "`", "{", "|", "}", "~", // easier to list these than calculate a formula
-}
-
-// sanitiseForTSVector escapes characters that have special meaning to ::TSVECTOR
-func sanitiseForTSVector(s string) string {
-	s = strings.ReplaceAll(s, `'`, `''`)
-	s = strings.ReplaceAll(s, `\`, `\\`)
-	return s
-}
-
-// desanitiseFromTSVector un-escapes characters that have special meaning to ::TSVECTOR
-// It is the inverse of sanitiseForTSVector
-func desanitiseFromTSVector(s string) string {
-	s = strings.ReplaceAll(s, `''`, `'`)
-	s = strings.ReplaceAll(s, `\\`, `\`)
-	return s
-}
-
-// keywordsToTSVector converts a slice of keywords to a
-// PostgreSQL tsvector. If the resulting tsvector would
-// be considered invalid by PostgreSQL an error is
-// returned instead.
-// `sep` SHOULD be either " " or "&". If "&", the output
-// string is a tsquery rather than a tsvector.
-func keywordsToTSVector(keywords []string, sep string) (string, error) {
-	const (
-		lexemeLimit   = 2048            // 2KB for single lexeme
-		tsvectorLimit = 1 * 1024 * 1024 // 1MB for lexemes + positions
-	)
-	newKeywords := []string{}
-	for _, k := range keywords {
-		if l := len(k); l >= lexemeLimit {
-			return "", fmt.Errorf("keyword exceeds limit (%d >= %d)", l, lexemeLimit)
-		}
-		newKeywords = append(newKeywords, fmt.Sprintf("'%s'", sanitiseForTSVector(k)))
-	}
-	tsv := strings.Join(newKeywords, sep)
-
-	if l := len(tsv); l >= tsvectorLimit {
-		return "", fmt.Errorf("keywords exceeds limit (%d >= %d)", l, tsvectorLimit)
-	}
-	return tsv, nil
-}
-
-// keywordsFromTSVector converts a PostgreSQL tsvector back into a slice of tokens.
-func keywordsFromTSVector(tsv string) (result []string) {
-	m := make(map[string]bool)
-	var s string
-	for {
-		i := strings.Index(tsv, "'")
-		if i == -1 {
-			break
-		}
-		tsv = tsv[i+1:]
-		i = strings.Index(tsv, "'")
-		if i == -1 {
-			break
-		}
-		s, tsv = tsv[:i], tsv[i+1:]
-		m[s] = true
-	}
-	for k := range m {
-		result = append(result, desanitiseFromTSVector(k))
-	}
-	return
-}
-
-// keywordsFromKey returns slices of keyword tokens, email addresses, and UIDs
-// extracted from the UserID packets of the given key.
-//
-// TODO: shouldn't this be a method on openpgp.PrimaryKey instead?
-// It's not specific to PostgreSQL, or even to storage.
-func keywordsFromKey(key *openpgp.PrimaryKey) (keywords []string, emails []string, uids []string) {
-	keywordMap := make(map[string]bool)
-	emailMap := make(map[string]bool)
-	uidMap := make(map[string]bool)
-	for _, uid := range key.UserIDs {
-		s := strings.ToLower(uid.Keywords)
-		// always include full text of UserID (lowercased)
-		keywordMap[s] = true
-		uidMap[s] = true
-		email := ""
-		commentary := s
-		lbr, rbr := strings.Index(s, "<"), strings.LastIndex(s, ">")
-		if lbr != -1 && rbr > lbr {
-			email = s[lbr+1 : rbr]
-			commentary = s[:lbr]
-		} else {
-			email = s
-			commentary = ""
-		}
-		// TODO: this still doesn't recognise all possible forms of UID :confounded:
-		if email != "" {
-			keywordMap[email] = true
-			emailMap[email] = true
-			parts := strings.SplitN(email, "@", 2)
-			if len(parts) == 2 {
-				keywordMap[parts[0]] = true
-				keywordMap[parts[1]] = true
-			}
-		}
-		if commentary != "" {
-			for _, field := range strings.FieldsFunc(commentary, func(r rune) bool {
-				return !utf8.ValidRune(r) || // split on invalid runes
-					!(unicode.IsLetter(r) || unicode.IsNumber(r) || r == '-' || r == '@') // split on [^[:alnum:]@-]
-			}) {
-				keywordMap[field] = true
-				for _, part := range strings.Split(field, "-") {
-					keywordMap[part] = true
-				}
-			}
-		}
-	}
-	for k := range keywordMap {
-		// discard empty strings, low ASCII symbols, single digits, stop words
-		if k == "" || (len(k) == 1 && k[0] < 0x41) || slices.Contains(pgEnglishStopWords, k) {
-			continue
-		}
-		keywords = append(keywords, k)
-	}
-	for k := range emailMap {
-		if k == "" {
-			continue
-		}
-		emails = append(emails, k)
-	}
-	for k := range uidMap {
-		if k == "" {
-			continue
-		}
-		uids = append(uids, k)
-	}
-	return
-}
-
-// keywordsFromSearch returns slices of keyword tokens and email addresses
-// extracted from the supplied search string.
-//
-// TODO: shouldn't this also be generic?
-func keywordsFromSearch(search string) (keywords []string, emails []string) {
-	keywordMap := make(map[string]bool)
-	emailMap := make(map[string]bool)
-	s := strings.ToLower(search)
-	email := s
-	lbr, rbr := strings.Index(s, "<"), strings.LastIndex(s, ">")
-	if lbr != -1 && rbr > lbr {
-		email = s[lbr+1 : rbr]
-		keywordMap[email] = true
-		emailMap[email] = true
-	} else {
-		for _, field := range strings.FieldsFunc(s, func(r rune) bool {
-			return !utf8.ValidRune(r) || unicode.IsSpace(r) // split on invalid runes and whitespace
-		}) {
-			keywordMap[field] = true
-		}
-	}
-	for k := range keywordMap {
-		// discard empty strings, low ASCII symbols, single digits, stop words
-		if k == "" || (len(k) == 1 && k[0] < 0x41) || slices.Contains(pgEnglishStopWords, k) {
-			continue
-		}
-		keywords = append(keywords, k)
-	}
-	for k := range emailMap {
-		if k == "" {
-			continue
-		}
-		emails = append(emails, k)
-	}
-	return
-}
-
-func subkeys(key *openpgp.PrimaryKey) []string {
-	var result []string
-	for _, subkey := range key.SubKeys {
-		result = append(result, subkey.RFingerprint)
-	}
-	return result
 }
